Use printf instead of echo for xmessage help
In some shells—like bash—the entire help message is shown in a single
line because a newline is printed as a literal "\n" character when using
echo.  Some distributions[1] have /bin/sh linked to bash, and so new
users are likely to run into this at some point.

We could fix this by either removing show and explicitly adding an
escaped quotation symbol before and after the string, or by using printf
instead of echo.  The printf solution seems more portable[2] and so I
propose we go with that one.

[1]: https://github.com/NixOS/nixpkgs/blob/f5b9a25cdd21b4e45ab5f11c27b95dfe17384274/nixos/modules/config/shells-environment.nix#L129-L143
[2]: https://unix.stackexchange.com/questions/65803/why-is-printf-better-than-echo/#65819